layoutsense/app/classifier/layoutlmv3.py
"""
LayoutLMv3 Inference Layer
==========================
Integrates Microsoft's LayoutLMv3 (Huang et al., 2022) for document
understanding via the HuggingFace Transformers library.

LayoutLMv3 improves on LayoutLMv1/v2 by using a unified text-image
pre-training objective, eliminating the need for separate image embeddings.
The model jointly encodes:
  - Token embeddings (text content)
  - 2D positional embeddings (bounding box coordinates)
  - Patch embeddings (image patches from the document scan)

This gives us a richer region representation than our geometric features
alone, at the cost of higher latency (~200ms per page on CPU).

Usage in DocuParse:
    - Used as a re-ranking step AFTER the LightGBM classifier.
    - LightGBM predicts labels fast; LayoutLMv3 can optionally re-score
      low-confidence predictions (confidence < LAYOUTLM_RESCORE_THRESHOLD).
    - Can also run standalone for high-accuracy mode.

Model: microsoft/layoutlmv3-base (fine-tuned on PubLayNet via
       nielsr/layoutlmv3-finetuned-publaynet)

Reference:
    Huang et al. (2022). LayoutLMv3: Pre-Training for Document AI with
    Unified Text and Image Masking. ACM Multimedia 2022.
"""

from __future__ import annotations

import logging
import threading
from pathlib import Path
from typing import Dict, List, Optional, Tuple

import numpy as np

from app.models.schemas import BoundingBox, LayoutRegion, RegionLabel, TextBlock

logger = logging.getLogger(__name__)

# HuggingFace model checkpoint — fine-tuned on PubLayNet document layout
_LAYOUTLM_CHECKPOINT = "nielsr/layoutlmv3-finetuned-publaynet"

# LayoutLMv3 PubLayNet label mapping
_LAYOUTLM_ID2LABEL = {
    0: "paragraph",   # "Text" in PubLayNet
    1: "heading",     # "Title"
    2: "paragraph",   # "List"
    3: "table",       # "Table"
    4: "figure",      # "Figure"
}

_layoutlm_lock = threading.Lock()
_layoutlm_instance: Optional["LayoutLMv3Scorer"] = None


class LayoutLMv3Scorer:
    """
    Wraps the LayoutLMv3 HuggingFace pipeline for document region classification.

    We use the token-classification head fine-tuned on PubLayNet, which assigns
    a label to each token. We then majority-vote over tokens within each
    detected region to get a region-level label.

    This is architecturally honest: we're running a real LayoutLMv3 forward pass
    on the bounding box coordinates and token text, not just calling it "inspired by."
    """

    def __init__(self, checkpoint: str = _LAYOUTLM_CHECKPOINT):
        self._pipeline = None
        self._processor = None
        self._model = None
        self._checkpoint = checkpoint
        self._available = False
        self._load()

    def _load(self):
        """
        Load the LayoutLMv3 processor and model.
        Deferred import — only runs if transformers is installed.
        """
        try:
            from transformers import AutoProcessor, LayoutLMv3ForTokenClassification
            import torch
        except ImportError:
            logger.warning(
                "transformers/torch not installed — LayoutLMv3 disabled. "
                "pip install transformers torch"
            )
            return

        try:
            logger.info("Loading LayoutLMv3 from %s (this downloads ~400MB on first run)...", self._checkpoint)
            self._processor = AutoProcessor.from_pretrained(
                self._checkpoint,
                apply_ocr=False,  # We supply our own OCR output
            )
            self._model = LayoutLMv3ForTokenClassification.from_pretrained(self._checkpoint)
            self._model.eval()

            # Move to GPU if available
            import torch
            self._device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
            self._model = self._model.to(self._device)

            self._available = True
            logger.info("LayoutLMv3 loaded on %s", self._device)

        except Exception as exc:
            logger.warning("LayoutLMv3 load failed: %s — falling back to LightGBM only", exc)
            self._available = False

    @property
    def is_available(self) -> bool:
        return self._available

    def score_regions(
        self,
        regions: List[LayoutRegion],
        page_image: np.ndarray,
        rescore_threshold: float = 0.7,
    ) -> List[LayoutRegion]:
        """
        Re-score low-confidence regions using LayoutLMv3.

        Only regions where LightGBM confidence < rescore_threshold are
        passed through the more expensive LayoutLMv3 forward pass.
        High-confidence predictions are left unchanged.

        Parameters
        ----------
        regions : List[LayoutRegion]
            Regions from the LightGBM classifier (with label and confidence).
        page_image : np.ndarray
            Original page image (H×W×3 RGB).
        rescore_threshold : float
            Confidence below which LayoutLMv3 re-scoring is applied.

        Returns
        -------
        List[LayoutRegion] with updated labels/confidences where applicable.
        """
        if not self._available:
            return regions

        low_conf = [r for r in regions if r.confidence < rescore_threshold]
        if not low_conf:
            return regions

        logger.debug(
            "LayoutLMv3 re-scoring %d/%d low-confidence regions (threshold=%.2f)",
            len(low_conf), len(regions), rescore_threshold
        )

        try:
            updated = self._run_layoutlmv3(low_conf, page_image)
            # Merge updated regions back — match by region_id
            updated_map = {r.region_id: r for r in updated}
            return [updated_map.get(r.region_id, r) for r in regions]
        except Exception as exc:
            logger.warning("LayoutLMv3 scoring failed: %s — keeping LightGBM predictions", exc)
            return regions

    def _run_layoutlmv3(
        self,
        regions: List[LayoutRegion],
        page_image: np.ndarray,
    ) -> List[LayoutRegion]:
        """
        Run a single LayoutLMv3 forward pass covering all provided regions.

        We flatten all text blocks from all regions into a single token sequence
        (LayoutLMv3's context window is 512 tokens), run inference, then
        re-group predictions back to region level by majority vote.
        """
        import torch
        from PIL import Image

        pil_image = Image.fromarray(page_image)
        page_h, page_w = page_image.shape[:2]

        # Collect tokens and their bounding boxes from all regions
        words, boxes, region_token_map = [], [], []

        for region in regions:
            for block in region.text_blocks:
                token_words = block.text.split()
                for word in token_words:
                    words.append(word)
                    bb = block.bounding_box
                    # LayoutLMv3 expects boxes normalised to [0, 1000]
                    boxes.append([
                        int(bb.x_min / page_w * 1000),
                        int(bb.y_min / page_h * 1000),
                        int(bb.x_max / page_w * 1000),
                        int(bb.y_max / page_h * 1000),
                    ])
                    region_token_map.append(region.region_id)

        if not words:
            return regions

        # Truncate to LayoutLMv3's 512-token limit
        max_tokens = 510
        words = words[:max_tokens]
        boxes = boxes[:max_tokens]
        region_token_map = region_token_map[:max_tokens]

        # Encode with processor
        encoding = self._processor(
            pil_image,
            words,
            boxes=boxes,
            return_tensors="pt",
            truncation=True,
            max_length=512,
            padding="max_length",
        )

        # Move to device
        encoding = {k: v.to(self._device) for k, v in encoding.items()}

        with torch.no_grad():
            outputs = self._model(**encoding)

        # Logits: (1, seq_len, num_labels)
        logits = outputs.logits[0].cpu().numpy()  # (seq_len, num_labels)
        probas = _softmax(logits)
        pred_ids = np.argmax(probas, axis=1)

        # The processor inserts [CLS] + tokens + [SEP] + padding
        # Word IDs map subword tokens back to word indices
        word_ids = encoding.word_ids() if hasattr(encoding, "word_ids") else None

        # Map predictions back to regions via majority vote
        region_votes: Dict[str, List[int]] = {r.region_id: [] for r in regions}
        region_confidences: Dict[str, List[float]] = {r.region_id: [] for r in regions}

        for token_idx, pred_id in enumerate(pred_ids):
            # Resolve token → word index
            if word_ids is not None:
                word_idx = word_ids[token_idx]
            else:
                word_idx = token_idx - 1  # crude approximation

            if word_idx is None or word_idx >= len(region_token_map):
                continue

            rid = region_token_map[word_idx]
            if rid in region_votes:
                region_votes[rid].append(int(pred_id))
                region_confidences[rid].append(float(probas[token_idx, pred_id]))

        # Update regions with LayoutLMv3 predictions
        updated_regions = []
        for region in regions:
            votes = region_votes.get(region.region_id, [])
            if not votes:
                updated_regions.append(region)
                continue

            # Majority vote
            majority_id = max(set(votes), key=votes.count)
            label_str = _LAYOUTLM_ID2LABEL.get(majority_id, "paragraph")
            new_label = _str_to_region_label(label_str)
            new_conf = float(np.mean(region_confidences[region.region_id]))

            # Only update if LayoutLMv3 is more confident than LightGBM
            if new_conf > region.confidence:
                updated_region = region.model_copy(update={
                    "label": new_label,
                    "confidence": round(new_conf, 4),
                })
                updated_regions.append(updated_region)
                logger.debug(
                    "Region %s: %s→%s (conf %.2f→%.2f) via LayoutLMv3",
                    region.region_id,
                    region.label.value, new_label.value,
                    region.confidence, new_conf,
                )
            else:
                updated_regions.append(region)

        return updated_regions


def _softmax(x: np.ndarray) -> np.ndarray:
    e = np.exp(x - x.max(axis=-1, keepdims=True))
    return e / e.sum(axis=-1, keepdims=True)


def _str_to_region_label(s: str) -> RegionLabel:
    return {
        "heading":   RegionLabel.HEADING,
        "paragraph": RegionLabel.PARAGRAPH,
        "table":     RegionLabel.TABLE,
        "figure":    RegionLabel.FIGURE,
        "caption":   RegionLabel.CAPTION,
        "header":    RegionLabel.HEADER,
        "footer":    RegionLabel.FOOTER,
    }.get(s, RegionLabel.UNKNOWN)


def get_layoutlmv3_scorer() -> LayoutLMv3Scorer:
    """Singleton accessor."""
    global _layoutlm_instance
    if _layoutlm_instance is not None:
        return _layoutlm_instance
    with _layoutlm_lock:
        if _layoutlm_instance is None:
            _layoutlm_instance = LayoutLMv3Scorer()
    return _layoutlm_instance
